requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
pdfplumber==0.9.0
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
python-pptx==0.6.21
Pillow==10.0.1
PyMuPDF==1.23.5